Fixed Created Sheets

+++ src/backend/migrations/2024-05-31-223058_create_publishers/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE publishers
 (
-    id      BIGINT PRIMARY KEY,
+    id      INTEGER PRIMARY KEY,
     username VARCHAR(100) NOT NULL,
     password VARCHAR(100) NOT NULL
 );
